refactor code in DeniedUsersAction component
